refactor: файловая структура sdk
