+++ package.json
@@ -12,8 +12,8 @@
         "prepublishOnly": "npm run build",
         "release": "auto shipit",
         "proto": "pbjs -t static-module src/proto/index.proto > src/proto/index.js && pbts src/proto/index.js -o src/proto/index.d.ts",
-        "asr": "pbjs -t static-module src/asr/index.proto > src/asr/index.js && pbts src/asr/index.js -o src/asr/index.d.ts",
-        "mtt": "pbjs -t static-module src/mtt/index.proto > src/mtt/index.js && pbts src/mtt/index.js -o src/mtt/index.d.ts",
+        "asr": "pbjs -t static-module src/assistantSdk/voice/asr/index.proto > src/assistantSdk/voice/asr/index.js && pbts src/assistantSdk/voice/asr/index.js -o src/assistantSdk/voice/asr/index.d.ts",
+        "mtt": "pbjs -t static-module src/assistantSdk/voice/mtt/index.proto > src/assistantSdk/voice/mtt/index.js && pbts src/assistantSdk/voice/mtt/index.js -o src/assistantSdk/voice/mtt/index.d.ts",
         "cy:open": "cypress open",
         "cy:run": "cypress run",
         "test:cy": "cypress run -b chromium --headless",
