+++ src/assistantSdk/voice/createNavigatorAudioProvider.ts
